# Eigen backend optimization for Dense, GRU and LSTM layers