… RTNeural/lstm/lstm_eigen.h
… RTNeural/lstm/lstm_eigen.tpp