Hello,

Recently, I read an [article on optimising LSTMs](https://developer.nvidia.com/blog/optimizing-recurrent-neural-networks-cudnn-5/) and was looking over the RTNeural implementations. This PR applies some of the ideas in the article to the Eigen backend.

Essentially, since Eigen vectorizes operations, it's advantageous to do fewer operations on larger operands, than many operations on smaller ones.

I only modified the dense, GRU and LSTM layers because those are the ones I'm familiar with.

## Benchmark results
I ran the benchmark with 10s signal, 16 in and 16 out, with the Release configuration, on a Macbook Air (2022, M2 16GB) and an Intel Core i5 11th gen with 32GB RAM running Windows 11.

Numbers are "times faster than real-time".

### M2 (Apple Clang)

| Layer | main | this branch |
| :--- | :--- | :--- |
| Dense (dynamic) | 414.80 | 693.66 |
| Dense (templated) | 720.45 | 826.06 |
| GRU (dynamic) | 47.53 | 67.16 |
| GRU (templated) | 59.83 | 68.43 |
| LSTM (dynamic) | 33.62 | 46.53 |
| LSTM (templated) | 45.10 | 49.18 |

### Intel (No AVX2, MSVC)

| Layer | main | this branch |
| :--- | :--- | :--- |
| Dense (dynamic) | 170.89 | 262.39 |
| Dense (templated) | 295.52 | 292.04 |
| GRU (dynamic) | 19.22 | 27.83 |
| GRU (templated) | 27.37 | 28.00 |
| LSTM (dynamic) | 13.50 | 20.50 |
| LSTM (templated) | 20.20 | 21.23 |

### Intel (No AVX2, Clang-cl)

| Layer | main | this branch |
| :--- | :--- | :--- |
| Dense (dynamic) | 188.74 | 305.48 |
| Dense (templated) | 314.58 | 327.81 |
| GRU (dynamic) | 21.12 | 29.46 |
| GRU (templated) | 29.07 | 31.44 |
| LSTM (dynamic) | 14.82 | 20.97 |
| LSTM (templated) | 20.71 | 22.86 |

### Intel (with AVX2, MSVC)

| Layer | main | this branch |
| :--- | :--- | :--- |
| Dense (dynamic) | 206.34 | 350.06 |
| Dense (templated) | 355.337 | 344.14 |
| GRU (dynamic) | 24.51 | 40.77 |
| GRU (templated) | 37.54 | 44.16 |
| LSTM (dynamic) | 17.76 | 30.95 |
| LSTM (templated) | 28.12 | 32.62 |

### Intel (with AVX2, Clang-cl)

| Layer | main | this branch |
| :--- | :--- | :--- |
| Dense (dynamic) | 217.82 | 400.09 |
| Dense (templated) | 433.96 | 470.44 |
| GRU (dynamic) | 27.96 | 48.41 |
| GRU (templated) | 45.27 | 50.16 |
| LSTM (dynamic) | 19.84 | 34.03 |
| LSTM (templated) | 33.09 | 37.35 |

I did not try it on a desktop Linux or a Raspberry Pi.

## Note on readability
The current implementation on main for the Eigen backend is extremely easy to read and for the most part one could probably just look it and understand the operations and their order in these layers (this is what I did about a year ago).

This new implementation, on the other hand, is quite difficult to read, I think. I left comments that hopefully explain what's going on, but they're too fresh in my eyes to know if they're useful. Naming suggestions and feedback for the comments would be *very very* helpful!

Thanks for reading this and hopefully it's useful!