@jatinchowdhury18 2023-08-07
Thanks for the pull request! I'm still going over all the changes, but they look great at first glance.

The failing tests are related to the sample-rate correction modes for the recurrent layers. I should be able to have a closer look at those tomorrow, but if you get to them first that's fine too :)

The performance numbers look fantastic. I've been using the XSIMD backend for most recurrent networks, but I should test again with the new changes (from the CI runs, XSIMD appears to still be a little bit faster, but benchmarks on the CI runners aren't very reliable). I did notice that the Dense templated layer appears to be a little bit slower on the Intel CPU with the new changes... The difference is small enough that I think the win on ARM is worth doing the change, but I'd like to test at a few more layer sizes first.

I wouldn't worry too much about readability for now... I would still consider the STL implementation quite readable, and the comments that you left with your changes look very helpful as well.