@IHorvalds 2023-08-07
Fixed it. `rtneural_tests all` passed on my machine.