@IHorvalds 2023-08-08
I compiled it with clang-cl on Windows with AVX2, Release config and this is the performance I got for the dense layer:
![image](https://github.com/jatinchowdhury18/RTNeural/assets/22452178/68289985-3b75-4dd2-9d6c-9ed66e087e7b)

Much better than MSVC somehow, but I really don't understand why...

I've also run it with a profiler to see where most of the time is being spent and the templated implementation doesn't appear in the hot path.
The dynamic implementation seems to spend quite a bit of time doing the copy though:
![image](https://github.com/jatinchowdhury18/RTNeural/assets/22452178/7c04ec38-afdb-4deb-bbaa-29cb8e57a031)


I'll come back to this this weekend, re-run the benchmarks with clang-cl and update the description of the PR.

I agree with your comments @jatinchowdhury18, it really would be nice to not have to copy the inputs/outputs into the extended arrays, but I can't come up with an elegant way to do it right now. 
Also, strangely enough the fastest way I measured to do the copies was actually doing the for-loops. `memcpy`, `std::copy` with iterators and with pointers and doing things like `ins_internal.segment(0, in_sizet) = ...` were measurably slower, haha.