@jatinchowdhury18 2023-08-09
> I compiled it with clang-cl on Windows with AVX2, Release config and this is the performance I got for the dense layer...
> Much better than MSVC somehow, but I really don't understand why...

This is really interesting! I have seen several cases where clang-cl does a better job of optimizing some code than MSVC, but the difference in this case is a bit more extreme than I would expect. (As an aside, I've been using clang-cl more and more lately, but that's mainly because I like being able to use the same compiler across Windows/Mac/Linux.)

> I've also run it with a profiler to see where most of the time is being spent and the templated implementation doesn't appear in the hot path. The dynamic implementation seems to spend quite a bit of time doing the copy though: ![image](https://user-images.githubusercontent.com/22452178/259233476-7c04ec38-afdb-4deb-bbaa-29cb8e57a031.png)

For the templated implementation the compiler is usually able to inline the entire `forward()` method for each layer into whatever context is calling it, so it's usually hard to get direct profiling information (sometimes it's possible to coax some information out of the generated assembly, but that can be tricky). Even for the dynamic implementation, I don't always trust which lines of code the profiler claims is eating up all the time. In this case, I would imagine that the `memcpy` and the `for` loop copies are both taking up 0.49%, and the matrix multiply is using 5.42% + 4.93%, but for some reason the profiler is splitting this up across two lines.

> I'll come back to this this weekend, re-run the benchmarks with clang-cl and update the description of the PR.

Sounds good! I'll leave the PR open for you to merge whenever you like. Since I've approved the PR you should have permission to merge it, but please let me know if that's not the case.

> I agree with your comments @jatinchowdhury18, it really would be nice to not have to copy the inputs/outputs into the extended arrays, but I can't come up with an elegant way to do it right now.

For now I think it's fine to merge these changes as-is, and then we can think more about how to remove some of the copy steps later on (unless some inspiration strikes one of us in the coming days).

> Also, strangely enough the fastest way I measured to do the copies for the input vector was actually doing the for-loops. `memcpy`, `std::copy` with iterators and with pointers and doing things like `ins_internal.segment(0, in_sizet) = ...` were measurably slower.

This is interesting... I'm curious how much the difference ended up being?