@IHorvalds 2023-08-09
The difference in performance for the dense layer on M2, 10s 16in 16out:
| | Dynamic | Templated |
| :--- | :--- | :--- |
| for-loop | 707.72x | 874.091x |
| memcpy | 658.057x | 861.577x |
| std::copy with pointers | 693.393x | 864.482x |
| std::copy with iterators | Unstable (between 588x - 660x) <br/> (using std::span) | 847.775x <br/> (also unstable, but not quite as much) |
| .segment<in_sizet>(0) = ... | - | 862.968x |

This is just for the input vector.

I realize it's slightly faster than the description of PR. These numbers go down a bit when the laptop heats up (passive cooling...), but the proportions keep.