@IHorvalds 2023-08-12
Hey @jatinchowdhury18, the last 2 commits revoked the permission to merge. When you have time, could you approve it again?
PS. I updated the description with the clang-cl benchmarks :)