@IHorvalds 2023-08-14
It still doesn't let me merge
<img width="855" alt="image" src="https://github.com/jatinchowdhury18/RTNeural/assets/22452178/45f4bc09-569c-4119-9dfc-b91a369a4b71">